refactor: 관심사 분리를 위한 컴포넌트 및 hook 분리
