+++ src/features/auth/AuthContext.ts
@@ -1,0 +1,23 @@
+import { createContext, useContext } from "react";
+
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+export interface AuthContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(
+  undefined,
+);
+
+export const useAuthContext = () => {
+  const context = useContext(AuthContext);
+  if (!context) throw new Error("AuthContext 밖에서 사용할 수 없습니다");
+  return context;
+};
