+++ src/features/notification/NotificationProvider.tsx
@@ -1,0 +1,33 @@
+import React, { useState, useCallback, useMemo } from "react";
+import { type Notification, NotificationContext } from "./NotificationContext";
+
+const NotificationProvider: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = useCallback(
+    (message: string, type: Notification["type"]) => {
+      const id = Date.now();
+      setNotifications((prev) => [...prev, { id, message, type }]);
+    },
+    [],
+  );
+
+  const removeNotification = useCallback((id: number) => {
+    setNotifications((prev) => prev.filter((n) => n.id !== id));
+  }, []);
+
+  const value = useMemo(
+    () => ({ notifications, addNotification, removeNotification }),
+    [notifications, addNotification, removeNotification],
+  );
+
+  return (
+    <NotificationContext.Provider value={value}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
+
+export default NotificationProvider;
