+++ src/features/theme/ThemeProvider.tsx
@@ -1,0 +1,20 @@
+import React, { useState, useMemo, useCallback } from "react";
+import { ThemeContext } from "./ThemeContext";
+
+const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
+  const [theme, setTheme] = useState("light");
+
+  const toggleTheme = useCallback(() => {
+    setTheme((prev) => (prev === "light" ? "dark" : "light"));
+  }, []);
+
+  const value = useMemo(() => ({ theme, toggleTheme }), [theme, toggleTheme]);
+
+  return (
+    <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>
+  );
+};
+
+export default ThemeProvider;
